feat: Add support for while loop in Boom language (#master)
This commit introduces support for while loop in the Boom language. This feature has been implemented in the 'while.boom' program and the Boom grammar (Boom.g4).
The changes include the addition of new constants for boolean values in 'CompileBoom.java', refactoring of exit and enter label handling in loops, and modification of certain condition handlers. Other changes include syntax cleanup, such as removal of excess newlines in 'while.boom'.